+++ inst/doc/vtreatCrossFrames.R
@@ -79,7 +79,7 @@ dTestTreated <- vtreat::prepare(treatments,dTest,
 dTest$predM1 <- predict(m1,newdata=dTestTreated,type='response')
 plotRes(dTest,'predM1','y','model1 on test')
 
-## ----crossframes---------------------------------------------------------
+## ----crossframes------------------------------------------------------------
 dTrain <- d[d$rgroup!='test',,drop=FALSE]
 dTest <- d[d$rgroup=='test',,drop=FALSE]
 prep <- vtreat::mkCrossFrameCExperiment(dTrain,
@@ -99,7 +99,7 @@ newvars <- sort(union(newvars,c("xBad1_catB","xBad2_catB","xBad3_catB")))
 print(newvars)
 dTrainTreated <- prep$crossFrame
 
-## ----xframemodel---------------------------------------------------------
+## ----xframemodel------------------------------------------------------------
 m1 <- glm(paste('y',paste(newvars,collapse=' + '),sep=' ~ '),
           data=dTrainTreated,family=binomial(link='logit'))
 print(summary(m1))